feat(CombineGWASWorkflowAttritionTable): added SVG icons
