+++ src/Analysis/GWASV2/Shared/AttritionTable/ChartIcons/BarChart.jsx
@@ -1,8 +1,15 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 
-const BarChart = ({})=> {
-    return (<h1>LOL</h1>);
-};
+const BarChart = () => {
+  return (
+
+  <svg xmlns="http://www.w3.org/2000/svg" style={{margin:"0 auto"}} width="16px" viewBox="0 0 64 64">
+  <g id="Group_550" data-name="Group 550" transform="translate(-1 -1)">
+    <path id="Path_4799" data-name="Path 4799" d="M15.545,42.091H3.909A2.909,2.909,0,0,1,1,39.182V15.909A2.909,2.909,0,0,1,3.909,13H15.545a2.909,2.909,0,0,1,2.909,2.909V39.182A2.909,2.909,0,0,1,15.545,42.091Z" transform="translate(0 22.909)" fill="#a7ccef"/>
+    <path id="Path_4800" data-name="Path 4800" d="M23.545,65H11.909A2.909,2.909,0,0,1,9,62.091V3.909A2.909,2.909,0,0,1,11.909,1H23.545a2.909,2.909,0,0,1,2.909,2.909V62.091A2.909,2.909,0,0,1,23.545,65Z" transform="translate(15.273 0)" fill="#2e77b8"/>
+    <path id="Path_4801" data-name="Path 4801" d="M31.545,49.727H19.909A2.909,2.909,0,0,1,17,46.818V11.909A2.909,2.909,0,0,1,19.909,9H31.545a2.909,2.909,0,0,1,2.909,2.909V46.818A2.909,2.909,0,0,1,31.545,49.727Z" transform="translate(30.545 15.273)" fill="#5495d0"/>
+  </g>
+</svg>);
+}
 
 export default BarChart;
