+++ src/Analysis/GWASV2/Shared/AttritionTable/ChartIcons/EulerDiagram.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const BarChart = () => {
+  return (
+
+<svg xmlns="http://www.w3.org/2000/svg" width="22px" viewBox="0 0 64 42.667">
+  <path id="Path_4817" data-name="Path 4817" d="M43.182,2.8A19.8,19.8,0,0,0,32.106,6.061,20.876,20.876,0,0,0,21.03,2.8,21.114,21.114,0,0,0,.106,24.13,21.114,21.114,0,0,0,21.03,45.464,19.8,19.8,0,0,0,32.106,42.2a20.876,20.876,0,0,0,11.076,3.264A21.114,21.114,0,0,0,64.106,24.13,21.114,21.114,0,0,0,43.182,2.8ZM32.106,10.7A17.008,17.008,0,0,1,34.2,12.834H30.015A17.29,17.29,0,0,1,32.106,10.7ZM27.059,30.4h9.97A16.051,16.051,0,0,1,35.8,32.913H28.169c-.37-.75-.74-1.632-1.11-2.509Zm-.736-10.037H37.768c.125.754.245,1.632.37,2.509H25.953C26.078,22,26.2,21.121,26.323,20.366Zm-.245,5.019H38.263a8.272,8.272,0,0,1-.37,2.509H26.444a8.273,8.273,0,0,1-.366-2.509Zm11.076-7.528h-9.97a16.051,16.051,0,0,1,1.231-2.509h7.629c.37.75.74,1.632,1.11,2.509ZM21.03,41.7a17.569,17.569,0,0,1,0-35.13,16.728,16.728,0,0,1,7.878,2.009,21.53,21.53,0,0,0,0,31.116,16.724,16.724,0,0,1-7.878,2Zm11.076-4.137a17.007,17.007,0,0,1-2.091-2.132H34.2a17.323,17.323,0,0,1-2.1,2.132Z" transform="translate(-0.106 -2.797)" fill="#2e77b8"/>
+</svg>);
+}
+
+export default BarChart;
