+++ src/Analysis/GWASV2/Shared/constants.js
@@ -17,70 +17,3 @@ export const gwasV2Steps = [
     secondaryTitle: 'Configure GWAS',
   },
 ];
-
-export const attritionTableHeaderConfig = [
-  {
-    title: 'Type',
-    dataIndex: 'type',
-    key: 'type',
-    align: 'center',
-  },
-  /*  {
-    title: 'Chart',
-    dataIndex: 'chart',
-    key: 'chart',
-    align: 'center',
-  }, */
-  {
-    title: 'Name',
-    dataIndex: 'name',
-    key: 'name',
-    align: 'center',
-  },
-  {
-    title: 'Size',
-    dataIndex: 'size',
-    key: 'size',
-    align: 'center',
-  },
-  {
-    title: 'Non-Hispanic Black',
-    dataIndex: 'nonHispanicBlack',
-    key: 'nonHispanicBlack',
-    align: 'center',
-    width: '18%',
-  },
-  {
-    title: 'Non-Hispanic Asian',
-    dataIndex: 'nonHispanicAsian',
-    key: 'nonHispanicAsian',
-    align: 'center',
-    width: '18%',
-  },
-  {
-    title: 'Non-Hispanic White',
-    dataIndex: 'nonHispanicWhite',
-    key: 'nonHispanicWhite',
-    align: 'center',
-    width: '18%',
-  },
-  {
-    title: 'Hispanic',
-    dataIndex: 'hispanic',
-    key: 'hispanic',
-    align: 'center',
-    width: '18%',
-  },
-];
-
-export const headerDataSource = [
-  {
-    type: '',
-    name: '',
-    size: '',
-    nonHispanicBlack: '',
-    nonHispanicAsian: '',
-    nonHispanicWhite: '',
-    hispanic: '',
-  },
-];
